# [ASM] Optimize body parsing - part 1