--- a/tracer/src/Datadog.Trace/AppSec/ObjectExtractor.cs
+++ b/tracer/src/Datadog.Trace/AppSec/ObjectExtractor.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reflection;
+using System.Reflection.Emit;
 using Datadog.Trace.AppSec.Waf;
 using Datadog.Trace.Logging;
 
@@ -18,6 +19,9 @@ namespace Datadog.Trace.AppSec
     {
         private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor(typeof(ObjectExtractor));
         private static readonly IReadOnlyDictionary<string, object> EmptyDictionary = new ReadOnlyDictionary<string, object>(new Dictionary<string, object>(0));
+
+        private static readonly object Sync = new();
+        private static readonly Dictionary<Type, FieldExtractor[]> TypeToExtractorMap = new();
 
         private static readonly HashSet<Type> AdditionalPrimitives = new()
         {
@@ -31,8 +35,8 @@ namespace Datadog.Trace.AppSec
 
         internal static object Extract(object body)
         {
-            var visted = new HashSet<object>();
-            var item = ExtractType(body.GetType(), body, 0, visted);
+            var visited = new HashSet<object>();
+            var item = ExtractType(body.GetType(), body, 0, visited);
 
             return item;
         }
@@ -53,42 +57,76 @@ namespace Datadog.Trace.AppSec
 
             Log.Debug("ExtractProperties - body: {Body}", body);
 
-            var fields = body.GetType()
-                        .GetFields(BindingFlags.Instance | BindingFlags.NonPublic)
-                        .Where(x => x.IsPrivate && x.Name.EndsWith("__BackingField"))
-                        .ToArray();
+            var bodyType = body.GetType();
 
             depth++;
 
-            var dictSize = Math.Min(WafConstants.MaxContainerSize, fields.Length);
+            FieldExtractor[] fieldExtractors = null;
+            lock (Sync)
+            {
+                if (!TypeToExtractorMap.TryGetValue(bodyType, out fieldExtractors))
+                {
+                    var fields =
+                        bodyType
+                           .GetFields(BindingFlags.Instance | BindingFlags.NonPublic)
+                           .Where(x => x.IsPrivate && x.Name.EndsWith("__BackingField"))
+                           .ToArray();
+
+                    fieldExtractors = new FieldExtractor[fields.Length];
+                    for (var i = 0; i < fields.Length; i++)
+                    {
+                        var field = fields[i];
+
+                        var propertyName = GetPropertyName(field.Name);
+                        if (string.IsNullOrEmpty(propertyName))
+                        {
+                            Log.Warning("ExtractProperties - couldn't extract property name from: {FieldName}", field.Name);
+                            continue;
+                        }
+
+                        var dynMethod = new DynamicMethod(
+                            bodyType + "_get_" + propertyName,
+                            typeof(object),
+                            new[] { typeof(object) },
+                            typeof(ObjectExtractor).Module,
+                            true);
+                        var ilGen = dynMethod.GetILGenerator();
+                        ilGen.Emit(OpCodes.Ldarg_0);
+                        ilGen.Emit(OpCodes.Ldfld, field);
+                        ilGen.Emit(OpCodes.Ret);
+                        var func = (Func<object, object>)dynMethod.CreateDelegate(typeof(Func<object, object>));
+
+                        fieldExtractors[i] = new FieldExtractor() { Name = propertyName, Type = field.FieldType, Accessor = func };
+                    }
+
+                    TypeToExtractorMap.Add(bodyType, fieldExtractors);
+                }
+            }
+
+            var dictSize = Math.Min(WafConstants.MaxContainerSize, fieldExtractors.Length);
             var dict = new Dictionary<string, object>(dictSize);
 
-            for (var i = 0; i < fields.Length; i++)
+            for (var i = 0; i < fieldExtractors.Length; i++)
             {
                 if (dict.Count >= WafConstants.MaxContainerSize || depth >= WafConstants.MaxContainerDepth)
                 {
                     return dict;
                 }
 
-                var field = fields[i];
+                var fieldExtractor = fieldExtractors[i];
 
-                var propertyName = GetPropertyName(field.Name);
-                if (string.IsNullOrEmpty(propertyName))
+                if (fieldExtractor != null)
                 {
-                    Log.Warning("ExtractProperties - couldn't extract property name from: {FieldName}", field.Name);
-                    continue;
+                    var value = fieldExtractor.Accessor.Invoke(body);
+                    Log.Debug("ExtractProperties - property: {Name} {Value}", fieldExtractor.Name, value);
+
+                    var item =
+                        value == null ?
+                            null :
+                            ExtractType(fieldExtractor.Type, value, depth, visited);
+
+                    dict.Add(fieldExtractor.Name, item);
                 }
-
-                var value = field.GetValue(body);
-
-                Log.Debug("ExtractProperties - property: {Name} {Value}", propertyName, value);
-
-                var item =
-                    value == null ?
-                        null :
-                        ExtractType(field.FieldType, value, depth, visited);
-
-                dict.Add(propertyName, item);
             }
 
             return dict;
@@ -187,6 +225,15 @@ namespace Datadog.Trace.AppSec
             }
 
             return items;
+        }
+
+        private class FieldExtractor
+        {
+#pragma warning disable CS0649, SA1401
+            public string Name;
+            public Type Type;
+            public Func<object, object> Accessor;
+#pragma warning restore CS0649, SA1401
         }
     }
 }