## Summary of changes

Replace reflection when reading properties with dynamic methods that can read the fields directly.

## Reason for change

Before this change:

|                   Method |          Mean |       Error |        StdDev |        Median |    Gen0 |  Allocated |
|------------------------- |--------------:|------------:|--------------:|--------------:|--------:|-----------:|
|   ObjectExtractorProps10 |      2.598 us |   0.0393 us |     0.0646 us |      2.581 us |  0.0076 |    1.45 KB |
|  ObjectExtractorProps100 |     23.830 us |   0.7047 us |     2.0333 us |     22.988 us |  0.0610 |    9.78 KB |
| ObjectExtractorProps1000 |    108.955 us |   2.9134 us |     8.3121 us |    106.032 us |  0.2441 |    42.2 KB |
| ObjectExtractorProps10x3 |     26.936 us |   0.6214 us |     1.7930 us |     26.200 us |  0.0916 |   16.02 KB |
| ObjectExtractorProps10x6 | 35,651.743 us | 645.4764 us | 1,163.9272 us | 35,252.567 us | 83.3333 | 21166.3 KB |

After:

|                   Method |           Mean |         Error |       StdDev |         Median |    Gen0 |    Gen1 | Allocated |
|------------------------- |---------------:|--------------:|-------------:|---------------:|--------:|--------:|----------:|
|   ObjectExtractorProps10 |       0.939 us |      0.038 us |     0.113 us |       0.888 us |  0.0038 |       - |     616 B |
|  ObjectExtractorProps100 |       8.051 us |      0.159 us |     0.369 us |       7.919 us |  0.0153 |       - |    3304 B |
| ObjectExtractorProps1000 |      20.303 us |      0.404 us |     0.604 us |      20.379 us |  0.0610 |       - |    8512 B |
| ObjectExtractorProps10x3 |       6.824 us |      0.161 us |     0.469 us |       6.661 us |  0.0381 |       - |    6040 B |
| ObjectExtractorProps10x6 |   9,725.468 us |    189.205 us |   239.283 us |   9,739.478 us | 46.8750 | 31.2500 | 7759260 B |


## Test coverage

Added new benchmarks (that are not run by the CI). This should be a transparent refactoring that already covered by unit tests and integration tests.

## Other details

This PR only looks at optimizations to reading objects with properties. I'll look if there are further optimizations in reading collections in another PR.